refactor: TDET-51 Reorganize DAG files to improve maintainability
